Remove dependency from ndt-server and reorganize model pkg.
