+++ pkg/ndtm/model/namevalue.go
@@ -1,0 +1,7 @@
+package model
+
+// NameValue is a BigQuery-compatible type for name/value pairs.
+type NameValue struct {
+	Name  string
+	Value string
+}
